Added onto the State Machine
I added in states for shielding as well as shooting for when we get the archer and shielder in. I also created two animator controllers that are used within the state scripts: KnightAnimatorController and ArcherAnimator controller. I tried to go into the relevant state scripts for them and start setting up code to help transition the animations when we plug them in