Added a  new way for the troop to move towards the target castle
Instead of using A* for the troop movement which would be a lot more complicated than we could figure out, I thought using raycasts would be much more simple. As of right now, the troop spawns in with a ray cast, and waits about 3 seconds before moving towards the enemy castle. I put this in the Knight script.